refactor: 체스 판을 출력하기 위해 메세지로 변환하는 책임을 DTO에서 View로 변경
