+++ src/test/java/chess/model/game/ChessGameTest.java
@@ -5,7 +5,7 @@ import static chess.helper.PositionFixture.A2;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.view.ChessBoardResponse;
+import chess.view.dto.ChessBoardResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
